Implemented "Keep running" option for schedules
If your script is running for a longer time, e.g. because it is running
in a long loop doing things and you like it to be restarted if it
ends/crashes for some reason, you can use the keep running option to
make the manager restart it for you.