Fixes pull request #5
Some method comments had to be altered. Plus I had to change the
dict to a special class to keep the interface like this:

for device in ccu.devices:
    for channel in device.channels:
        ...

With just a simple dict it had to be

for device in ccu.devices:
    for channel in device.channels.values():
        ...

And in this case the channels would not be sorted by channel index.